Consolidate DB Writer code into one shared class
