+++ f/connectors/geojson/geojson_to_postgres.py
@@ -5,9 +5,7 @@ import json
 import logging
 from pathlib import Path
 
-from psycopg2 import connect, errors, sql
-
-from f.common_logic.db_operations import conninfo, postgresql
+from f.common_logic.db_operations import FormDBWriter, conninfo, postgresql
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -23,7 +21,13 @@ def main(
     geojson_path = Path(attachment_root) / Path(geojson_path)
     transformed_geojson_data = transform_geojson_data(geojson_path)
 
-    db_writer = GeoJSONDbWriter(conninfo(db), db_table_name)
+    db_writer = FormDBWriter(
+        conninfo(db),
+        db_table_name,
+        use_mapping_table=False,
+        sanitize_keys=True,
+        reverse_properties_separated_by=None,
+    )
     db_writer.handle_output(transformed_geojson_data)
 
     if delete_geojson_file:
@@ -59,218 +63,6 @@ def transform_geojson_data(geojson_path):
         }
         transformed_geojson_data.append(transformed_feature)
     return transformed_geojson_data
-
-
-class GeoJSONDbWriter:
-    """
-    Converts GeoJSON spatial data to structured SQL tables.
-    """
-
-    def __init__(self, db_connection_string, table_name):
-        """
-        Initializes the GeoJSONIOManager with the provided connection string and form response table to be used.
-        """
-        self.db_connection_string = db_connection_string
-        self.table_name = table_name
-
-    def _get_conn(self):
-        """
-        Establishes a connection to the PostgreSQL database using the class's configured connection string.
-        """
-        return connect(dsn=self.db_connection_string)
-
-    def _inspect_schema(self, table_name):
-        """
-        Fetches the column names of the given table.
-        """
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        cursor.execute(
-            "SELECT column_name FROM information_schema.columns WHERE table_name = %s",
-            (table_name,),
-        )
-        columns = [row[0] for row in cursor.fetchall()]
-        cursor.close()
-        conn.close()
-        return columns
-
-    def _create_missing_fields(self, table_name, missing_columns):
-        """
-        Generates and executes SQL statements to add missing fields to the table.
-        """
-        table_name = sql.Identifier(table_name)
-        try:
-            with self._get_conn() as conn, conn.cursor() as cursor:
-                query = sql.SQL(
-                    "CREATE TABLE IF NOT EXISTS {table_name} (_id TEXT PRIMARY KEY);"
-                ).format(table_name=table_name)
-                cursor.execute(query)
-
-                for sanitized_column in missing_columns:
-                    if sanitized_column == "_id":
-                        continue
-                    try:
-                        query = sql.SQL(
-                            "ALTER TABLE {table_name} ADD COLUMN {colname} TEXT;"
-                        ).format(
-                            table_name=table_name,
-                            colname=sql.Identifier(sanitized_column),
-                        )
-                        cursor.execute(query)
-                    except errors.DuplicateColumn:
-                        logger.debug(
-                            f"Skipping insert due to DuplicateColumn, this form column has been accounted for already in the past: {sanitized_column}"
-                        )
-                        continue
-                    except Exception as e:
-                        logger.error(
-                            f"An error occurred while creating missing column: {sanitized_column} for {table_name}: {e}"
-                        )
-                        raise
-        finally:
-            conn.close()
-
-    @staticmethod
-    def _safe_insert(cursor, table_name, columns, values):
-        """
-        Executes a safe INSERT operation into a PostgreSQL table, ensuring data integrity and preventing SQL injection.
-        This method also handles conflicts by updating existing records if necessary.
-
-        The function first checks if a row with the same primary key (_id) already exists in the table. If it does,
-        and the existing row's data matches the new values, the operation is skipped. Otherwise, it performs an
-        INSERT operation. If a conflict on the primary key occurs, it updates the existing row with the new values.
-
-        Parameters
-        ----------
-        cursor : psycopg2 cursor
-            The database cursor used to execute SQL queries.
-        table_name : str
-            The name of the table where data will be inserted.
-        columns : list of str
-            The list of column names corresponding to the values being inserted.
-        values : list
-            The list of values to be inserted into the table, aligned with the columns.
-
-        Returns
-        -------
-        tuple
-            A tuple containing two integers: the count of rows inserted and the count of rows updated.
-        """
-        inserted_count = 0
-        updated_count = 0
-
-        # Check if there is an existing row that is different from the new values
-        # We are doing this in order to keep track of which rows are actually updated
-        # (Otherwise all existing rows would be added to updated_count)
-        id_index = columns.index("_id")
-        values[id_index] = str(values[id_index])
-        select_query = sql.SQL("SELECT {fields} FROM {table} WHERE _id = %s").format(
-            fields=sql.SQL(", ").join(map(sql.Identifier, columns)),
-            table=sql.Identifier(table_name),
-        )
-        cursor.execute(select_query, (values[columns.index("_id")],))
-        existing_row = cursor.fetchone()
-
-        if existing_row and list(existing_row) == values:
-            # No changes, skip the update
-            return inserted_count, updated_count
-
-        query = sql.SQL(
-            "INSERT INTO {table} ({fields}) VALUES ({placeholders}) "
-            "ON CONFLICT (_id) DO UPDATE SET {updates} "
-            # The RETURNING clause is used to determine if the row was inserted or updated.
-            # xmax is a system column in PostgreSQL that stores the transaction ID of the deleting transaction.
-            # If xmax is 0, it means the row was newly inserted and not updated.
-            "RETURNING (xmax = 0) AS inserted"
-        ).format(
-            table=sql.Identifier(table_name),
-            fields=sql.SQL(", ").join(map(sql.Identifier, columns)),
-            placeholders=sql.SQL(", ").join(sql.Placeholder() for _ in values),
-            updates=sql.SQL(", ").join(
-                sql.Composed(
-                    [sql.Identifier(col), sql.SQL(" = EXCLUDED."), sql.Identifier(col)]
-                )
-                for col in columns
-                if col != "_id"
-            ),
-        )
-
-        cursor.execute(query, values)
-        result = cursor.fetchone()
-        if result and result[0]:
-            inserted_count += 1
-        else:
-            updated_count += 1
-
-        return inserted_count, updated_count
-
-    def handle_output(self, outputs):
-        """
-        Inserts GeojSON spatial data into the specified PostgreSQL database table.
-        It checks the database schema and adds any missing fields, then constructs
-        and executes SQL insert queries to store the data. After processing all data,
-        it commits the transaction and closes the database connection.
-        """
-        table_name = self.table_name
-
-        conn = self._get_conn()
-        cursor = conn.cursor()
-
-        # Safely truncate the table to 63 characters
-        # TODO: ...while retaining uniqueness
-        table_name = table_name[:63]
-
-        existing_fields = self._inspect_schema(table_name)
-        rows = []
-        for entry in outputs:
-            sanitized_entry = {k: v for k, v in entry.items()}
-            rows.append(sanitized_entry)
-
-        missing_field_keys = set()
-        for row in rows:
-            missing_field_keys.update(set(row.keys()).difference(existing_fields))
-
-        if missing_field_keys:
-            self._create_missing_fields(table_name, missing_field_keys)
-
-        logger.info(
-            f"Attempting to write {len(rows)} submissions to the DB to table {table_name}"
-        )
-
-        inserted_count = 0
-        updated_count = 0
-
-        for row in rows:
-            try:
-                cols, vals = zip(*row.items())
-
-                # Serialize lists, dict values to JSON text
-                vals = list(vals)
-                for i in range(len(vals)):
-                    value = vals[i]
-                    if isinstance(value, list) or isinstance(value, dict):
-                        vals[i] = json.dumps(value)
-
-                result_inserted_count, result_updated_count = self._safe_insert(
-                    cursor, table_name, cols, vals
-                )
-                inserted_count += result_inserted_count
-                updated_count += result_updated_count
-            except Exception as e:
-                logger.error(f"Error inserting data: {e}, {type(e).__name__}")
-                conn.rollback()
-
-            try:
-                conn.commit()
-            except Exception as e:
-                logger.error(f"Error committing transaction: {e}")
-                conn.rollback()
-
-        logger.info(f"Total rows inserted: {inserted_count}")
-        logger.info(f"Total rows updated: {updated_count}")
-
-        cursor.close()
-        conn.close()
 
 
 def delete_geojson_file(
