+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 skipsdist = True
-env_list = alerts, arcgis, comapeo, geojson, globalforestwatch, kobotoolbox_responses, locusmap, odk_responses, postgres_to_file
+env_list = alerts, arcgis, common_logic, comapeo, geojson, globalforestwatch, kobotoolbox_responses, locusmap, odk_responses, postgres_to_file
 
 [testenv]
 setenv =
@@ -37,6 +37,14 @@ deps =
     -r{toxinidir}/f/connectors/comapeo/tests/requirements-test.txt
 commands =
     pytest {posargs} f/connectors/comapeo
+
+[testenv:common_logic]
+deps =
+    -r{toxinidir}/f/common_logic/db_transformations.script.lock
+    -r{toxinidir}/f/common_logic/tests/requirements-test.txt
+commands =
+    pytest {posargs} f/common_logic
+
 
 [testenv:geojson]
 deps =
